fix undeclared mapped types
This integration was previously creating mapped _types like the
following: `mapping_souce_has_azure_vm` or `mapping_source_is_azure_vm`.

This is undesired behavior since it will mark the data sets as partial.
This commit adds 6 new types relating `tenable_vulnerability_finding`
and `tennable_asset` to the instance for the three major CSPs.

Additionaly, it removes a mapped relationships that was relating
`tenable_assets` to `tenable_assets`. This seemed like a mistake to me.
If we intened to keep this relationship we should just make it a direct
relationship.

For `Hosts|Devices` that are not one of the three major CSP _types we
should instead rely on global mapping rules to map based on properties
like **ipv4**, **ipv6**, or **fqdn**. Checking that these rules are in
place is a separate task.

I've also removed an unused utility function normalizeCVSS2Severity.